pyproject.toml
[project]
name = "langgraph-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.10.0",
    "deepgram-sdk>=5.3.0",
    "duckduckgo-search>=8.1.1",
    "flashrank>=0.2.10",
    "fpdf>=1.7.2",
    "gradio==6.0.1",
    "grandalf>=0.8",
    "langchain>=1.0.8",
    "langchain-azure-ai>=1.0.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-tavily>=0.2.15",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "markdown>=3.10",
    "mermaid-python>=0.1",
    "opik>=1.9.46",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "tiktoken>=0.12.0",
    "unstructured>=0.18.21",
    "youtube-transcript-api>=1.2.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pre-commit>=4.5.1",
    "ruff>=0.14.8",
]

[project.scripts]
doc_gen = "graph_examples.doc_generator.doc_gen_app:main"
rag_search = "graph_examples.rag_search.rag_search_app:main"
product_review = "graph_examples.review_product.run_product_review:main"

[tool.ruff]
fix = true
unsafe-fixes = false
line-length = 88

[tool.ruff.lint]
select = [
    "F",  # Core errors like undefined names, unused imports, and unused variables.
    "E",  # PEP 8 style errors (e.g., incorrect indentation, missing whitespace).
    "W",  # PEP 8 style warnings (e.g., trailing whitespace, blank lines).
    "I",  # Ensures imports are sorted and grouped correctly (stdlib, 3rd party, local).
    "UP", # Suggests modern Python syntax (e.g., using f-strings instead of % formatting).
    "B",  # Catches common bugs and design problems (e.g., mutable default arguments).
    "Q",  # Enforces consistent quote style (single vs. double quotes).
    "N",  # Checks naming conventions (e.g., ClassName, function_name, CONSTANT).
]

ignore = [
    "B008", # Mutable default arguments.
    "D413", # Missing blank line after last section in docstring.
    "E501", # Line too long.
    "W293", # Blank line contains whitespace.
    "W291", # Trailing whitespace.
    "PLR2004", # Magic value used in comparison (e.g., if x == 5).
    "D101", # Missing docstring in public class.
    "D400", # First line should end with a period and be followed by a blank line.
]

fixable = ["ALL"] # All rules that have auto-fixes are allowed to be fixed automatically.

extend-select = [
    "PLR2004", # Magic values in comparisons (e.g., if status == 200). Encourages constants.
    "D101", # Missing docstring in a public class.
    "D102", # Missing docstring in a public method.
    "D103", # Missing docstring in a public function.
    "D201", # No blank lines allowed before function docstring.
    "D202", # No blank lines allowed after function docstring.
    "D215", # Section underline is over-indented in docstring.
    "T201", # print() found. Helps catch debug prints left in code.
]

[tool.mypy]
disallow_any_generics = true
disallow_untyped_defs = true
disallow_untyped_globals = true
ignore_missing_imports = true
show_error_codes = true
strict = false
warn_return_any = true
plugins = ['pydantic.mypy']
packages = ["graph_examples"]



